Validatorの基底クラスを作成
stringとimageのバリデーションで使用する共通処理を基底クラスに抽出しました。
